Lab 14b

+++ src/test/java/com/example/selenium/spree/ScreenshotRule.java
@@ -5,15 +5,23 @@ import org.junit.rules.TestWatcher;
 import org.junit.runner.Description;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.IOException;
 
 public class ScreenshotRule extends TestWatcher {
 
+  private static final org.slf4j.Logger log = LoggerFactory.getLogger(ScreenshotRule.class);
+
   private final SeleniumTest seleniumTest;
-  private final String screenShotDir;
+  private final File screenShotDir;
 
   public ScreenshotRule(SeleniumTest seleniumTest, String screenShotDir) {
+    this(seleniumTest, new File(screenShotDir));
+  }
+
+  public ScreenshotRule(SeleniumTest seleniumTest, File screenShotDir) {
     this.seleniumTest = seleniumTest;
     this.screenShotDir = screenShotDir;
   }
@@ -29,21 +37,43 @@ public class ScreenshotRule extends TestWatcher {
   @Override
   protected void failed(Throwable exception, Description description) {
     RemoteWebDriver webDriver = seleniumTest.getRemoteWebDriver();
-    File dir = new File(screenShotDir);
-    if (dir.exists() == false) dir.mkdirs();
+    if (webDriver == null) {
+      log.error("The web driver does not exist.");
+      return;
+    }
 
-    File srcFile = webDriver.getScreenshotAs(OutputType.FILE);
+    if (screenShotDir.exists() == false) {
+      if (screenShotDir.mkdirs() == false) {
+        log.error("Unable to create the working directory ({}) for capturing screen shots.", screenShotDir.getAbsolutePath());
+      }
+    }
+
+    File srcFile;
+
+    try {
+      srcFile = webDriver.getScreenshotAs(OutputType.FILE);
+    } catch (Exception ex) {
+      log.error("Unable to capture screen shot: {}", ex.getMessage());
+      return;
+    }
+
     int pos = srcFile.getName().lastIndexOf(".");
     String fileExt = srcFile.getName().substring(pos+1);
 
     String fileName = String.format("screen-shot-%s.%s", description.getDisplayName(), fileExt);
     File dstFile = new File(screenShotDir, fileName);
 
+    if (dstFile.exists() && dstFile.delete() == false) {
+      log.error("Unable to delete previous screen shot: {}", dstFile.getAbsolutePath());
+      return;
+    }
+
     try {
       FileUtils.copyFile(srcFile, dstFile);
-      System.out.println("Screen shot saved to " + dstFile.getAbsolutePath());
-    } catch (Exception ex) {
-      ex.printStackTrace();
+    } catch (IOException ioe) {
+      log.error("Unable to copy temporary screen shot:\nsrc: {}\ndst: {}", srcFile.getAbsolutePath(), dstFile.getAbsolutePath());
     }
+
+    log.warn("Screen shot saved to {}", dstFile.getAbsolutePath());
   }
 }
